update test program and a makefile Made changes to testwiringpi.c when replicating the lesson #3 from Osoyoo. In my previous doing of this lesson I must have made a mistake in the circuit so that the pin commands of HIGH and LOW were swapped.
However when redoing to experiment from scratch the program was incorrect
in that when the program thought it was turning the LED off it was instead
lighting the LED andwhen turning the LED on it was instead turning it off.

Also now have a make file to make the executable wiretest.