added GPIO #18 and -b option to change brightness of LED Now a potentiometer can be connected to a channel of the ADC and an LED connected through a resistor to GPIO #18 to allow changing the brightness of the LED by changing the voltage on a channel with the potentiometer.
The new option is -b and must be used with sudo.

so you can use a command line of "./spitest 3 -b" and the application
will check the third channel of the MCP3008 (CH2) and read the value
there and then change the brightness level of the LED to that setting.