source code for the photoresistor project starting with the application from project07 we have added an additional command line argument, -p, to allow the user to specify an ADC channel on the MCP3008 to which is attached a photoresistor connected to 3.3v with a 10K ohm pullup resistor.
The voltage value detected by the MCP3008 varies between 0v and
3.3v depending on the resistance offered by the photoresistor
whose resistance in turn varies depending on the amount of
light it receives. The more light then the lower the resistence
of the photoresistor and the more of the 3.3v rail voltage that
will be seen by the MCP3008.